Don't delete the old .jar is the plugin is persistant. Fixes issue 1333
git-svn-id: http://svn.dmdirc.com/trunk@4197 00569f92-eb28-0410-84fd-f71c24880f43

+++ src/com/dmdirc/updater/components/PluginComponent.java
@@ -86,7 +86,7 @@ public class PluginComponent implements UpdateComponent {
     public boolean doInstall(final String path) throws Throwable {
         final File target = new File(plugin.getFullFilename());
         
-        if (target.exists()) {
+        if (!plugin.hasPersistant() && target.exists()) {
             target.delete();
         }
